+++ addons/fix_PO/CfgMagazineWells.hpp
@@ -1,7 +1,0 @@
-class CfgMagazineWells {
-    class CBA_303B_LeeEn {
-        WB_Magazines[] += {
-            "LOP_10rnd_77mm_mag"
-        };
-    };
-};