+++ addons/fix_PO/CfgVehicles.hpp
@@ -1,6 +1,0 @@
-class CfgVehicles {
-    class B_Kitbag_cbr;
-    class LOP_B_KB_Med_tub: B_Kitbag_cbr{
-        displayName = "Kitbag (Tubitak/Medic)";
-    };
-};