+++ mod.cpp
@@ -1,6 +1,5 @@
-name = "Weapons balance - PO";
-description = "A collection of weight balancing mods made by johnb43";
-tooltip = "Weight balance";
-tooltipOwned = "Weight balance";
-overview = "";
+name = "Weapons Balance - Project Opfor";
+description = "A collection of balancing mods made by johnb43";
+tooltip = "Weapons Balance - Project Opfor";
+tooltipOwned = "Weapons Balance - Project Opfor";
 author = "johnb43";